REFACTOR: rearranging file and folder structure
