+++ src/camera/camera.c
@@ -34,6 +34,35 @@ t_camera	camera(double hsize, double vsize, double field_of_view)
 	return (c);
 }
 
+t_matrix	view_transform(t_point from, t_point to, t_vector up)
+{
+	t_vector	forward;
+	t_vector	left;
+	t_vector	true_up;
+	t_matrix	orientation;
+	t_matrix	view_matrix;
+
+	forward = normalize(tuple_subtract(to, from));
+	if (fmod(fabs(dot(forward, up)), 1) < EPSILON)
+		left = vector(-1, 0, 0);
+	else
+		left = cross(forward, normalize(up));
+	true_up = cross(left, forward);
+	orientation = identity();
+	mx_set(&orientation, 0, 0, left.x);
+	mx_set(&orientation, 0, 1, left.y);
+	mx_set(&orientation, 0, 2, left.z);
+	mx_set(&orientation, 1, 0, true_up.x);
+	mx_set(&orientation, 1, 1, true_up.y);
+	mx_set(&orientation, 1, 2, true_up.z);
+	mx_set(&orientation, 2, 0, -forward.x);
+	mx_set(&orientation, 2, 1, -forward.y);
+	mx_set(&orientation, 2, 2, -forward.z);
+	view_matrix = mx_multiply(
+			orientation, translation(-from.x, -from.y, -from.z));
+	return (view_matrix);
+}
+
 t_ray	ray_for_pixel(t_camera c, int x, int y, t_anti_aliasing aa_data)
 {
 	double		x_offset;
@@ -54,80 +83,4 @@ t_ray	ray_for_pixel(t_camera c, int x, int y, t_anti_aliasing aa_data)
 					origin
 				)))
 	);
-}
-
-static t_color	render_pixel(t_camera c, t_world w, int x, int y)
-{
-	t_anti_aliasing	aa_data;
-	t_color			color_average_result;
-	t_ray			ray;
-	t_color			*colors;
-
-	colors = malloc(sizeof(t_color) * pow(w.pixel_sampling, 2));
-	aa_data.pixel_sampling = w.pixel_sampling;
-	aa_data.sample_y = 0;
-	while (aa_data.sample_y < w.pixel_sampling)
-	{
-		aa_data.sample_x = 0;
-		while (aa_data.sample_x < w.pixel_sampling)
-		{
-			ray = ray_for_pixel(c, x, y, aa_data);
-			colors[(int) aa_data.sample_x + (int)(aa_data.sample_y
-					* w.pixel_sampling)] = color_at(w, ray, 4); // color_at ganhou novo argumento (depth),
-																// precisa verificar como preencher aqui
-																// setei 4 por recomendação do livro (default)
-																// Dá pra colocar em uma variável no t_world
-			aa_data.sample_x++;
-		}
-		aa_data.sample_y++;
-	}
-	color_average_result = color_average(colors, pow(w.pixel_sampling, 2));
-	free(colors);
-	return (color_average_result);
-}
-
-void	*render_line(void *param)
-{
-	int			x;
-	t_color		coloring;
-	t_thread	data;
-
-	data = (*(t_thread *)param);
-	x = 0;
-	while (x < data.line_size)
-	{
-		coloring = render_pixel(data.camera, data.world, x, data.line);
-		write_pixel_to_canvas(data.canvas, x, data.line, coloring);
-		x++;
-	}
-	return (NULL);
-}
-
-t_canvas	render(t_camera c, t_world w)
-{
-	int			y;
-	int			thread_count;
-	t_canvas	image;
-	t_thread	threads_data[NUM_THREADS];
-	pthread_t	threads[NUM_THREADS];
-
-	y = -1;
-	thread_count = 0;
-	image = create_canvas(c.hsize, c.vsize);
-	print_rendering_progress(c.hsize, c.vsize, 0);
-	while (++y < c.vsize)
-	{
-		if (thread_count < NUM_THREADS)
-		{
-			threads_data[thread_count] = (t_thread){.line = y, \
-		.line_size = c.hsize, .canvas = &image, .world = w, .camera = c};
-			pthread_create(&threads[thread_count], NULL,
-				render_line, &threads_data[thread_count]);
-			thread_count++;
-		}
-		thread_count = reset_threads(threads, thread_count);	// Uma pequena gambiarra
-		print_rendering_progress(c.hsize, c.vsize, y);
-	}
-	join_threads(threads, thread_count);
-	return (image);
 }
