+++ src/materials/material.c
@@ -12,18 +12,6 @@
 
 #include "minirt.h"
 
-t_pattern	default_pattern(void)
-{
-	return ((t_pattern){
-		.pattern_type = STRIPE,
-		.has_pattern = false,
-		.color_a = color(0, 0, 0),
-		.color_b = color(0, 0, 0),
-		.transform = identity(),
-		.inverse = identity()
-	});
-}
-
 /**
  * @brief Creates and returns a default material.
  *
