+++ src/materials/patterns/checkers.c
