+++ src/materials/patterns/gradient.c
