+++ src/materials/patterns/ring.c
