+++ src/materials/patterns/stripe.c
