+++ src/render/intersections/intersect_utils.c
